Make Hero Headline static (medium and large screens)
- Fixes text overflow when constantly re-sizing the screen

Optmizes for the important information for the user, I think, resulting
in a better user experience.

Also: the purpose of the portfolio is to introduce the developer (hi! that's me) and her work. This means the light headline background box will cover more of the background image for some screen sizes. And that is ok.

Previous iterations were more visually stressful when clicking and
dragging through every pixel of screen size difference because the text
size kept having to grow or shrink to fit the box. This was optimizing
for the background image